Solve failing tests while packaging with maven
Tests pass individualy, but when tested during test phase while packaging,
some tests fail because of @DynamicPropertySource. It is registrering
testcontainer uri only for first test and the rest will fail because of
new testcontainer instances created with new random exposed ports.
Add @DirtiesContext to make new ApplicationContext for each test class
that extends AbstractTestcontainersUnitTest and by that register proper
uri of testcontainer.

+++ src/test/java/dev/akibaz/movies/dao/MovieRepositoryTest.java
@@ -21,7 +21,9 @@ class MovieRepositoryTest extends AbstractTestcontainersUnitTest {
 
     @BeforeEach
     void setUp() {
-        underTest.deleteAll();
+        if (!underTest.findAll().isEmpty()) {
+            underTest.deleteAll();
+        }
     }
 
     @Test
